refactoring - make services, DTOs, repos
