+++ src/main/kotlin/com/example/blog/controller/HtmlController.kt
@@ -1,6 +1,10 @@
-package com.example.blog
+package com.example.blog.controller
 
-import org.springframework.http.HttpStatus
+import com.example.blog.BlogProperties
+import com.example.blog.format
+import com.example.blog.model.ArticleDTO.Article
+import com.example.blog.service.ArticleService
+import com.example.blog.service.UserService
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.ui.set
@@ -8,33 +12,32 @@ import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.server.ResponseStatusException
 
 @Controller
 class HtmlController(
-    private val articleRepository: ArticleRepository,
+    private val articleService: ArticleService,
+    private val userService: UserService,
     private val properties: BlogProperties,
-    private val userRepository: UserRepository,
 ) {
     @GetMapping("/", "/home")
     fun blog(model: Model): String {
         model["title"] = properties.title
         model["banner"] = properties.banner
-        model["articles"] = articleRepository.findAllByOrderByAddedAtDesc().map { it.render() }
+        model["articles"] = articleService.findAllArticles().map { it.render() }
         return "blog"
     }
 
     @GetMapping("/articles")
     fun getArticles(model: Model): String {
         model["banner"] = properties.banner
-        model["articles"] = articleRepository.findAllByOrderByAddedAtDesc().map { it.render() }
+        model["articles"] = articleService.findAllArticles().map { it.render() }
         return "articles"
     }
 
     @GetMapping("/users")
     fun getUsers(model: Model): String {
         model["banner"] = properties.banner
-        model["users"] = userRepository.findAll()
+        model["users"] = userService.findAllUsers()
         return "users"
     }
 
@@ -43,12 +46,7 @@ class HtmlController(
         @PathVariable slug: String,
         model: Model,
     ): String {
-        val article =
-            articleRepository
-                .findBySlug(slug)
-                ?.render()
-                ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "This article does not exist")
-        println("Rendering article: $article")
+        val article = articleService.findArticleBySlug(slug).render()
         model["banner"] = properties.banner
         model["title"] = article.title
         model["article"] = article
@@ -64,17 +62,7 @@ class HtmlController(
         @RequestParam content: String,
         @RequestParam username: String,
     ): String {
-        val user: User =
-            userRepository.findByUsername(username)
-                ?: throw IllegalArgumentException("Author not found")
-
-        val article =
-            Article(
-                title = title,
-                content = content,
-                author = user,
-            )
-        articleRepository.save(article)
+        articleService.createArticle(title, content, username)
         return "redirect:/articles"
     }
 
@@ -87,13 +75,7 @@ class HtmlController(
         @RequestParam firstname: String,
         @RequestParam lastname: String,
     ): String {
-        val user =
-            User(
-                username = username,
-                firstName = firstname,
-                lastName = lastname,
-            )
-        userRepository.save(user)
+        userService.createUser(username, firstname, lastname)
         return "redirect:/users"
     }
 
@@ -104,10 +86,7 @@ class HtmlController(
     fun deleteUser(
         @RequestParam username: String,
     ): String {
-        val user: User =
-            userRepository.findByUsername(username)
-                ?: throw IllegalArgumentException("User not found")
-        userRepository.delete(user)
+        userService.deleteUserByUsername(username)
         return "redirect:/users"
     }
 
@@ -118,10 +97,7 @@ class HtmlController(
     fun deleteArticle(
         @RequestParam title: String,
     ): String {
-        val article: Article =
-            articleRepository.findByTitleIgnoreCase(title)
-                ?: throw ResponseStatusException(HttpStatus.NOT_FOUND, "This article does not exist")
-        articleRepository.delete(article)
+        articleService.deleteArticleByTitle(title)
         return "redirect:/articles"
     }
