+++ src/main/kotlin/com/example/blog/repository/UserRepository.kt
@@ -1,0 +1,8 @@
+package com.example.blog.repository
+
+import com.example.blog.model.UserDTO.User
+import org.springframework.data.repository.CrudRepository
+
+interface UserRepository : CrudRepository<User, Long> {
+    fun findByUsername(username: String): User?
+}
